refactor(speccolor): rename the file to set color parameters
make it easy to identify which files provide the design specs, rather than css value calculations
